♻️: 1. refactored losses and re-added EnhancedCriterion  2. refactored experiments yaml files
