+++ configs/experiment/vpc/vpc_amm.yaml
@@ -1,0 +1,97 @@
+# @package _global_
+
+# to execute this experiment run:
+# python train.py experiment=example
+
+defaults:
+  - override /callbacks: default.yaml
+  - override /datamodule: datasets/vpc.yaml
+  - override /module: vpc.yaml
+  - override /trainer: gpu.yaml
+  - override /logger: tensorboard.yaml
+
+# all parameters below will be merged with parameters from default configurations set above
+# this allows you to overwrite only specified parameters
+
+tags: ["vpc25", "de-anonmyzation", "amm"]
+seed: 42
+
+trainer:
+  min_epochs: 10
+  max_epochs: 25
+  gradient_clip_val: 5.0
+  num_sanity_val_steps: 0
+
+callbacks:
+  model_checkpoint:
+    mode: min
+    monitor: ${replace:"valid/__metric__"}/${module.metrics.valid_best.target_key}
+  early_stopping:
+    mode: min
+    monitor: ${replace:"valid/__metric__"}/${module.metrics.valid_best.target_key}
+    patience: 10
+
+datamodule:
+  dataset:
+    max_duration: 12.0
+
+  models:
+    B3: ${datamodule.available_models.B3}
+    B5: ${datamodule.available_models.B5}
+    T8-5: ${datamodule.available_models.T8-5}
+
+  loaders:
+    train:
+      batch_size: 16
+    valid:
+      batch_size: 16
+    test:
+      batch_size: 16
+    enrollment:
+      batch_size: 1
+
+module:
+  normalize_test_scores: True
+  model:
+    classifiers: 
+      selected_classifier: robust   # normalized robust
+      
+  criterion:
+    selected_criterion: comprehensive   # aam  cross_entropy comprehensive
+    available_criteria:
+      comprehensive:
+        _target_: src.modules.losses.components.multi_modal_losses.MultiModalLoss
+        classifier_name: ${module.model.classifiers.selected_classifier}
+        weights:
+          _target_: src.modules.losses.components.multi_modal_losses.LossWeights
+          ensemble: ${map:${module.model.classifiers.selected_classifier},robust,1.0,0.0}
+          fusion: ${map:${module.model.classifiers.selected_classifier},normalized,1.0,0.2}
+          audio: 0.2
+          text: 0.2
+          contrastive: 0.1
+          consistency: 1.0
+          confidence: 0.1
+        confidence_target: 0.9
+        classification_loss: 
+          _target_: speechbrain.nnet.losses.LogSoftmaxWrapper
+          loss_fn:
+            _target_: speechbrain.nnet.losses.AdditiveAngularMargin
+            margin: 0.2
+            scale: 30.0
+
+  scheduler_settings:
+    selected_lr_scheduler: reduce_on_plateau  # reduce_on_plateau warmup_cosine
+    available_lr_scheduler:
+      reduce_on_plateau:
+        scheduler:
+          mode: "min"
+
+  optimizer:
+    lr: 1.0e-4
+    weight_decay: 1.0e-5
+
+
+# logger:
+#   wandb:
+#     tags: ${tags}
+#     group: "mnist"
