+++ configs/experiment/vpc/vpc_amm_aug.yaml
@@ -4,18 +4,16 @@
 # python train.py experiment=example
 
 defaults:
-  # - override /callbacks: default.yaml
-  - /callbacks/model_pruning
-  - override /datamodule: datasets/voxceleb.yaml
-  - override /module: sv.yaml
+  - override /callbacks: default.yaml
+  - override /datamodule: datasets/vpc.yaml
+  - override /module: vpc.yaml
   - override /trainer: gpu.yaml
   - override /logger: tensorboard.yaml
-
 
 # all parameters below will be merged with parameters from default configurations set above
 # this allows you to overwrite only specified parameters
 
-tags: ["data_augmentation", "model_pruning", "voxceleb", "sv", "ecapa"]
+tags: ["data_augmentation", "vpc25", "de-anonmyzation", "amm"]
 seed: 42
 
 trainer:
@@ -26,47 +24,77 @@ trainer:
 
 callbacks:
   model_checkpoint:
-    mode: max
-    monitor: ${replace:"valid/__metric__"}
+    mode: min
+    monitor: ${replace:"valid/__metric__"}/${module.metrics.valid_best.target_key}
   early_stopping:
-    mode: max
-    monitor: ${replace:"valid/__metric__"}
-    patience: 8
+    mode: min
+    monitor: ${replace:"valid/__metric__"}/${module.metrics.valid_best.target_key}
+    patience: 10
 
 datamodule:
+  dataset:
+    max_duration: 12.0
+
+  models:
+    B3: ${datamodule.available_models.B3}
+    B5: ${datamodule.available_models.B5}
+    T8-5: ${datamodule.available_models.T8-5}
+
   loaders:
     train:
-      batch_size: 16
+      batch_size: 4
     valid:
-      batch_size: 16
+      batch_size: 4
     test:
-      batch_size: 16
+      batch_size: 4
     enrollment:
       batch_size: 1
 
 module:
-  normalize_test_scores: True
 
+  normalize_test_scores: True
+  model:
+    classifiers: 
+      selected_classifier: robust   # normalized robust
+      
   criterion:
-    selected_criterion: aam   # aam  cross_entropy
+    selected_criterion: comprehensive   # aam  cross_entropy comprehensive
+    available_criteria:
+      comprehensive:
+        _target_: src.modules.losses.components.multi_modal_losses.MultiModalLoss
+        classifier_name: ${module.model.classifiers.selected_classifier}
+        weights:
+          _target_: src.modules.losses.components.multi_modal_losses.LossWeights
+          ensemble: ${map:${module.model.classifiers.selected_classifier},robust,1.0,0.0}
+          fusion: ${map:${module.model.classifiers.selected_classifier},normalized,1.0,0.2}
+          audio: 0.2
+          text: 0.2
+          contrastive: 0.1
+          consistency: 1.0
+          confidence: 0.1
+        confidence_target: 0.9
+        classification_loss: 
+          _target_: speechbrain.nnet.losses.LogSoftmaxWrapper
+          loss_fn:
+            _target_: speechbrain.nnet.losses.AdditiveAngularMargin
+            margin: 0.2
+            scale: 30.0
 
   scheduler_settings:
-    selected_lr_scheduler: warmup_cosine  # reduce_on_plateau warmup_cosine
+    selected_lr_scheduler: reduce_on_plateau  # reduce_on_plateau warmup_cosine
     available_lr_scheduler:
       reduce_on_plateau:
         scheduler:
-          mode: "max"
+          mode: "min"
 
   optimizer:
     lr: 1.0e-4
     weight_decay: 1.0e-5
 
-  ### Data for augmentation ###
   data_augemntation:
-
     noise_dataset_url: https://www.dropbox.com/scl/fi/a09pj97s5ifan81dqhi4n/noises.zip?rlkey=j8b0n9kdjdr32o1f06t0cw5b7&dl=1
     rir_dataset_url: https://www.dropbox.com/scl/fi/linhy77c36mu10965a836/RIRs.zip?rlkey=pg9cu8vrpn2u173vhiqyu743u&dl=1
-    data_folder_rir_noises: ${datamodule.dataset.data_dir}/RIRS_NOISES
+    data_folder_rir_noises: ${paths.data_dir}/voxceleb/RIRS_NOISES
     noise_annotation: ${module.data_augemntation.data_folder_rir_noises}/noise.csv
     rir_annotation: ${module.data_augemntation.data_folder_rir_noises}/reverb.csv
 
@@ -88,8 +116,8 @@ module:
       add_noise:
         _target_: speechbrain.augment.time_domain.AddNoise
         csv_file: ${module.data_augemntation.noise_annotation}
-        snr_low: 0
-        snr_high: 15
+        snr_low: 10
+        snr_high: 20
         noise_sample_rate: ${datamodule.dataset.sample_rate}
         clean_sample_rate: ${datamodule.dataset.sample_rate}
         num_workers: ${datamodule.loaders.train.num_workers}
@@ -111,8 +139,8 @@ module:
 
       drop_chunk:
         _target_: speechbrain.augment.time_domain.DropChunk
-        drop_length_low: 1000
-        drop_length_high: 2000
+        drop_length_low: 2000
+        drop_length_high: 8000
         drop_count_low: 1
         drop_count_high: 5
 
@@ -133,4 +161,3 @@ module:
 #   wandb:
 #     tags: ${tags}
 #     group: "mnist"
-
