+++ configs/experiment/vpc/vpc_ce.yaml
@@ -13,7 +13,7 @@ defaults:
 # all parameters below will be merged with parameters from default configurations set above
 # this allows you to overwrite only specified parameters
 
-tags: ["vpc25", "de-anonmyzation"]
+tags: ["vpc25", "de-anonmyzation", "CrossEntropy"]
 seed: 42
 
 trainer:
@@ -58,6 +58,22 @@ module:
       
   criterion:
     selected_criterion: comprehensive   # aam  cross_entropy comprehensive
+    available_criteria:
+      comprehensive:
+        _target_: src.modules.losses.components.multi_modal_losses.MultiModalLoss
+        classifier_name: ${module.model.classifiers.selected_classifier}
+        weights:
+          _target_: src.modules.losses.components.multi_modal_losses.LossWeights
+          ensemble: ${map:${module.model.classifiers.selected_classifier},robust,1.0,0.0}
+          fusion: ${map:${module.model.classifiers.selected_classifier},normalized,1.0,0.2}
+          audio: 0.2
+          text: 0.2
+          contrastive: 0.1
+          consistency: 1.0
+          confidence: 0.1
+        confidence_target: 0.9
+        classification_loss: 
+          _target_: torch.nn.CrossEntropyLoss
 
   scheduler_settings:
     selected_lr_scheduler: reduce_on_plateau  # reduce_on_plateau warmup_cosine
