+++ configs/module/vpc.yaml
@@ -3,75 +3,6 @@ _target_: src.modules.multimodal_vpc.MultiModalVPCModel
 defaults:
   - _self_
 
-### Data for augmentation ###
-# data_augemntation:
-
-#   noise_dataset_url: https://www.dropbox.com/scl/fi/a09pj97s5ifan81dqhi4n/noises.zip?rlkey=j8b0n9kdjdr32o1f06t0cw5b7&dl=1
-#   rir_dataset_url: https://www.dropbox.com/scl/fi/linhy77c36mu10965a836/RIRs.zip?rlkey=pg9cu8vrpn2u173vhiqyu743u&dl=1
-#   data_folder_rir_noises: ${paths.data_dir}/RIRS_NOISES
-#   noise_annotation: ${module.data_augemntation.data_folder_rir_noises}/noise.csv
-#   rir_annotation: ${module.data_augemntation.data_folder_rir_noises}/reverb.csv
-
-#   data:
-#     noise_dataset:
-#       _target_: speechbrain.augment.preparation.prepare_dataset_from_URL
-#       URL: ${module.data_augemntation.noise_dataset_url}
-#       dest_folder: ${module.data_augemntation.data_folder_rir_noises}
-#       ext: wav
-#       csv_file: ${module.data_augemntation.noise_annotation}
-
-#   augmentations:
-#     add_noise:
-#       _target_: speechbrain.augment.time_domain.AddNoise
-#       csv_file: ${module.data_augemntation.noise_annotation}
-#       snr_low: 0
-#       snr_high: 15
-#       noise_sample_rate: ${datamodule.dataset.sample_rate}
-#       clean_sample_rate: ${datamodule.dataset.sample_rate}
-#       num_workers: ${datamodule.loaders.train.num_workers}
-
-#     prepare_rir:
-#       _target_: speechbrain.augment.preparation.prepare_dataset_from_URL
-#       URL: ${module.data_augemntation.rir_dataset_url}
-#       dest_folder: ${module.data_augemntation.data_folder_rir_noises}
-#       ext: wav
-#       csv_file: ${module.data_augemntation.rir_annotation}
-
-#     add_reverb:
-#       _target_: speechbrain.augment.time_domain.AddReverb
-#       csv_file: ${module.data_augemntation.rir_annotation}
-#       reverb_sample_rate: ${datamodule.dataset.sample_rate}
-#       clean_sample_rate: ${datamodule.dataset.sample_rate}
-#       num_workers: ${datamodule.loaders.train.num_workers}
-
-#     drop_freq:
-#       _target_: speechbrain.augment.time_domain.DropFreq
-#       drop_freq_low: 0
-#       drop_freq_high: 1
-#       drop_freq_count_low: 1
-#       drop_freq_count_high: 3
-#       drop_freq_width: 0.05
-
-#     drop_chunk:
-#       _target_: speechbrain.augment.time_domain.DropChunk
-#       drop_length_low: 1000
-#       drop_length_high: 2000
-#       drop_count_low: 1
-#       drop_count_high: 5
-
-#     wav_augmenter:
-#       _target_: speechbrain.augment.augmenter.Augmenter
-#       parallel_augment: true
-#       concat_original: true
-#       min_augmentations: 4
-#       max_augmentations: 4
-#       augment_prob: 1.0
-#       augmentations:
-#         - ${module.data_augemntation.augmentations.add_noise}
-#         - ${module.data_augemntation.augmentations.add_reverb}
-#         - ${module.data_augemntation.augmentations.drop_freq}
-#         - ${module.data_augemntation.augmentations.drop_chunk}
-
 
 model:
   audio_processor:
@@ -152,18 +83,25 @@ model:
 ### criteria and metrics ###
 criterion:
   available_criteria:
-    _base_aam_: &_base_aam_
+    amm: &_aam_
       _target_: speechbrain.nnet.losses.LogSoftmaxWrapper
       loss_fn:
         _target_: speechbrain.nnet.losses.AdditiveAngularMargin
         margin: 0.2
         scale: 30.0
 
+    cross_entropy:  &_ce_
+      _target_: torch.nn.CrossEntropyLoss
+
+    enhanced:
+      _target_: src.modules.multimodal_vpc.EnhancedCriterion
+      classification_loss: ???
+
     comprehensive:
-      _target_: src.modules.multimodal_vpc.MultiModalLoss
+      _target_: src.modules.losses.components.multi_modal_losses.MultiModalLoss
       classifier_name: ${module.model.classifiers.selected_classifier}
       weights:
-        _target_: src.modules.multimodal_vpc.LossWeights
+        _target_: src.modules.losses.components.multi_modal_losses.LossWeights
         ensemble: ${map:${module.model.classifiers.selected_classifier},robust,1.0,0.0}
         fusion: ${map:${module.model.classifiers.selected_classifier},normalized,1.0,0.2}
         audio: 0.2
@@ -172,15 +110,8 @@ criterion:
         consistency: 1.0
         confidence: 0.1
       confidence_target: 0.9
-      classification_loss: 
-        <<: *_base_aam_
+      classification_loss: ???
     
-    aam:
-      <<: *_base_aam_
-    
-    cross_entropy:
-      _target_: torch.nn.CrossEntropyLoss
-
   # Selected criterion (defaults to comprehensive)
   selected_criterion: ???
   train_criterion: ${module.criterion.available_criteria.${module.criterion.selected_criterion}}
